feat: 검색 결과 위젯 구현
