+++ lib/modules/search/widget/search_result_widget.dart
@@ -1,0 +1,64 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:get/get.dart';
+import 'package:ibaji/modules/detail_method/view/detail_method_view.dart';
+import 'package:ibaji/provider/service/map_service.dart';
+import 'package:ibaji/util/app_text_styles.dart';
+
+import '../../../model/trash/trash.dart';
+import '../../../util/app_colors.dart';
+
+class SearchResult extends StatelessWidget {
+  final Trash trash;
+  const SearchResult({super.key, required this.trash});
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        GestureDetector(
+          onTap: () async {
+            // Get.put(DetailMethodController(id: trash.id));
+            await Future.delayed(const Duration(seconds: 2));
+            await Get.to(const DetailMethodScreen(),
+                arguments: {'id': trash.id});
+          },
+          child: Container(
+            decoration: BoxDecoration(
+              borderRadius: BorderRadius.circular(8.r),
+            ),
+            child: Row(children: [
+              Container(
+                width: 60.w,
+                height: 60.h,
+                decoration: BoxDecoration(
+                    color: AppColors.grey1,
+                    borderRadius: BorderRadius.circular(6.r)),
+                child: Image.asset("asset/image/icon/ic_paper_bag.png"),
+              ),
+              SizedBox(
+                width: 13.w,
+              ),
+              Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    "${MapService.currentAddress[1]} ${MapService.currentAddress[2]} 기준",
+                    style: AppTextStyles.body1Medium
+                        .copyWith(color: AppColors.grey5),
+                  ),
+                  Text(
+                    trash.name,
+                    style: AppTextStyles.title2Bold,
+                  ),
+                ],
+              ),
+            ]),
+          ),
+        ),
+      ],
+    );
+  }
+}
